requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==4.2.13
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
easy-thumbnails==2.8.5
idna==3.7
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.5.0
urllib3==2.2.1
